Added RTKQ mutations for booking ActivityRequests
Added a mutation:

To start negotiations, setting the status of the
ActivityRequest to 'negotiating'

To cancel/back-out-of negotiations, setting the status of the
ActivityRequest back to 'pending'

To book an activity, setting the status of the ActivityRequest to
'booked'.

NOTE: Both users involved with this request mutually agree on whether
the activity has been booked and on the date they have agreed upon.
Currently, it does not do this, there may need to be some more back-end
code written to deal with this.